feat: 사회재난 알림 바텀시트뷰 제작 & 모달 컴포넌트 분리
